+++ sanic/asgi.py
@@ -1,16 +1,19 @@
 import asyncio
 import warnings
+
 from inspect import isawaitable
 from typing import Any, Awaitable, Callable, MutableMapping, Optional, Union
 from urllib.parse import quote
 
 import sanic.app  # noqa
+
 from sanic.compat import Header
 from sanic.exceptions import InvalidUsage
 from sanic.request import Request
 from sanic.response import HTTPResponse, StreamingHTTPResponse
 from sanic.server import ConnInfo
 from sanic.websocket import WebSocketConnection
+
 
 ASGIScope = MutableMapping[str, Any]
 ASGIMessage = MutableMapping[str, Any]
@@ -54,7 +57,9 @@ class MockProtocol:
 class MockTransport:
     _protocol: Optional[MockProtocol]
 
-    def __init__(self, scope: ASGIScope, receive: ASGIReceive, send: ASGISend) -> None:
+    def __init__(
+        self, scope: ASGIScope, receive: ASGIReceive, send: ASGISend
+    ) -> None:
         self.scope = scope
         self._receive = receive
         self._send = send
@@ -132,7 +137,9 @@ class Lifespan:
         ) + self.asgi_app.sanic_app.listeners.get("after_server_start", [])
 
         for handler in listeners:
-            response = handler(self.asgi_app.sanic_app, self.asgi_app.sanic_app.loop)
+            response = handler(
+                self.asgi_app.sanic_app, self.asgi_app.sanic_app.loop
+            )
             if isawaitable(response):
                 await response
 
@@ -150,7 +157,9 @@ class Lifespan:
         ) + self.asgi_app.sanic_app.listeners.get("after_server_stop", [])
 
         for handler in listeners:
-            response = handler(self.asgi_app.sanic_app, self.asgi_app.sanic_app.loop)
+            response = handler(
+                self.asgi_app.sanic_app, self.asgi_app.sanic_app.loop
+            )
             if isawaitable(response):
                 await response
 
@@ -199,7 +208,11 @@ class ASGIApp:
         if scope["type"] == "lifespan":
             await instance.lifespan(scope, receive, send)
         else:
-            path = scope["path"][1:] if scope["path"].startswith("/") else scope["path"]
+            path = (
+                scope["path"][1:]
+                if scope["path"].startswith("/")
+                else scope["path"]
+            )
             url = "/".join([scope.get("root_path", ""), quote(path)])
             url_bytes = url.encode("latin-1")
             url_bytes += b"?" + scope["query_string"]
@@ -222,7 +235,12 @@ class ASGIApp:
 
             request_class = sanic_app.request_class or Request
             instance.request = request_class(
-                url_bytes, headers, version, method, instance.transport, sanic_app,
+                url_bytes,
+                headers,
+                version,
+                method,
+                instance.transport,
+                sanic_app,
             )
             instance.request.stream = instance
             instance.request_body = True  # FIXME: Use more_body?
@@ -270,6 +288,8 @@ class ASGIApp:
                 "more_body": not end_stream,
             }
         )
+
+    _asgi_single_callable = True  # We conform to ASGI 3.0 single-callable
 
     async def __call__(self) -> None:
         """