+++ sanic/compat.py
@@ -1,5 +1,6 @@
 import asyncio
 import signal
+
 from sys import argv
 
 from multidict import CIMultiDict  # type: ignore