+++ sanic/static.py
@@ -13,6 +13,7 @@ from sanic.exceptions import (
     InvalidUsage,
 )
 from sanic.handlers import ContentRangeHandler
+from sanic.log import error_logger
 from sanic.response import HTTPResponse, file, file_stream
 
 
@@ -40,6 +41,10 @@ async def _static_request_handler(
     # match filenames which got encoded (filenames with spaces etc)
     file_path = path.abspath(unquote(file_path))
     if not file_path.startswith(path.abspath(unquote(root_path))):
+        error_logger.exception(
+            f"File not found: path={file_or_directory}, "
+            f"relative_url={file_uri}"
+        )
         raise FileNotFound(
             "File not found", path=file_or_directory, relative_url=file_uri
         )
@@ -94,6 +99,10 @@ async def _static_request_handler(
     except ContentRangeError:
         raise
     except Exception:
+        error_logger.exception(
+            f"File not found: path={file_or_directory}, "
+            f"relative_url={file_uri}"
+        )
         raise FileNotFound(
             "File not found", path=file_or_directory, relative_url=file_uri
         )
@@ -134,6 +143,8 @@ def register(
                               threshold size to switch to file_stream()
     :param name: user defined name used for url_for
     :param content_type: user defined content type for header
+    :return: registered static routes
+    :rtype: List[sanic.router.Route]
     """
     # If we're not trying to match a file directly,
     # serve from the folder
@@ -155,10 +166,11 @@ def register(
         )
     )
 
-    app.route(
+    _routes, _ = app.route(
         uri,
         methods=["GET", "HEAD"],
         name=name,
         host=host,
         strict_slashes=strict_slashes,
     )(_handler)
+    return _routes